Added basic damage calculation
Also moved a couple of files and made player scripts to use the same core scripts as enemies.
Actor states also now use strings and are filled up during runtime.

The hit/damage detaction is still very very basic. There is a lot more to do for a proper system, but for now it gets the job done.