Combat Scripts update
- Attack Core States updated
- AttackCore export vars now have min/max values
- Removed unneeded export vars in AttackCore
- Added factions to ActorCore
- AttackCore now checks for factions before colliding
- renamed colliding check function (in case we want to use collision shapes for the attack itself for things like bouncing off a wall)
- fixed a couple of minor rotation/positioning issues

I temporarily disabled the players attack's aiming towards the crosshair in order to test my script.

There are currently three factions: Player, Allies, Enemies
Each one of them cannot hurt actors within the same group, but can hurt the other two factions. Allies are supposed to be "NPCs" that arent enemies, but I'm not sure if we are going to need them. I just put them already in place.
The reason why allies can be attacked and attack player factions is that
1) if we dont want that to be, we can just give them the player faction
2) it was easier to make the checks within the script
3) i think it would be more interesting. you could "accidentally" (or purposely) kill npcs or they might attack you if you piss them off.